Update DragAndDropGrid.as
When dragging a column header and the horizontal scroll bar becomes active, the dragDropIndicator becomes out of line with the edges of the column headers. This is because the coordinates do not take account of the scroll position. Therefore I have added a new variable called scrollPosition and used it in several of the coordinates of the function showColumnDropFeedback.